src/requirements.txt
rospkg_modules==1.3.0
Pillow==8.3.2
rospkg==1.3.0
